# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BID \u2014 Batch Image Delivery API",
    "description": "REST API for the **BID** (*Batch Image Delivery*) image-processing pipeline.\n\nBID automates batch photo processing: automatic scaling, watermark overlay, EXIF cleaning, colour-space conversion, and delivery-folder export.  This API exposes the processing core as an OpenAPI 3.0-compliant service, decoupled from the legacy Tkinter UI.\n\n## Road-map\n| Phase | Milestone | Status |\n|-------|-----------|--------|\n| 1 | Backend API Extraction | **In progress** |\n| 2 | WebSocket real-time layer | Planned |\n| 3 | React/TypeScript frontend | Planned |\n| 4 | Event system | Planned |\n\n## Auth\nProtected endpoints require `Authorization: Bearer <access_token>`.  Obtain tokens via `POST /api/v1/auth/login`.  Admin-only endpoints additionally require the `admin` role.\n\n## Errors\nAll error responses use the `ErrorResponse` schema: `{ \"detail\": \"...\", \"field\": \"...\" }`.",
    "contact": {
      "name": "BID project",
      "url": "https://github.com/bajdero/BID"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login and receive JWT tokens",
        "description": "Authenticate with *username* and *password*.\n\nOn success returns an **access token** (short-lived) and a **refresh token**\n(long-lived).  Pass the access token as `Authorization: Bearer <token>` on\nprotected endpoints.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh access token",
        "description": "Exchange a valid *refresh_token* for a new access + refresh token pair.\n\nPreviously issued refresh tokens remain valid until they expire according to\nthe server's token lifetime settings; clients should discard old refresh\ntokens once a new pair is obtained.\n\n# TODO(security): implement refresh-token rotation \u2014 store JTIs in a DB\n# revocation table and invalidate the previous token on re-issue.\n# See: https://github.com/bajdero/BID/pull/26#discussion (C4)",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/exports/conflicts": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "List blocked / missing export files",
        "description": "Return export entries where the expected output file is missing or has\nzero bytes on disk \u2014 indicating the photo needs to be re-processed.",
        "operationId": "list_conflicts_api_v1_projects__project_id__exports_conflicts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConflictItem"
                  },
                  "title": "Response List Conflicts Api V1 Projects  Project Id  Exports Conflicts Get"
                }
              }
            }
          },
          "404": {
            "description": "Project not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/exports/conflicts/resolve": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Resolve export conflicts",
        "description": "Apply *body.action* to the matching conflict records.\n\n- ``action=\"replace\"`` resets state to ``new`` (re-queues for processing).\n- ``action=\"skip\"``    marks the photo as ``skip``.",
        "operationId": "resolve_conflicts_api_v1_projects__project_id__exports_conflicts_resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictResolutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Number of resolved conflicts and the action applied.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Resolve Conflicts Api V1 Projects  Project Id  Exports Conflicts Resolve Post"
                }
              }
            }
          },
          "404": {
            "description": "Project not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictItem": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "folder": {
            "type": "string",
            "title": "Folder"
          },
          "photo": {
            "type": "string",
            "title": "Photo"
          },
          "target_path": {
            "type": "string",
            "title": "Target Path"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "missing_or_empty"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "folder",
          "photo",
          "target_path"
        ],
        "title": "ConflictItem",
        "description": "A single blocked / missing export entry.",
        "example": {
          "folder": "Session1",
          "photo": "IMG_001.jpg",
          "profile": "fb",
          "reason": "missing_or_empty",
          "target_path": "fb/Session1/YAPA2026-03-23_Session1_IMG_001.jpg"
        }
      },
      "ConflictResolutionRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "all",
              "selection"
            ],
            "title": "Mode"
          },
          "action": {
            "type": "string",
            "enum": [
              "replace",
              "skip"
            ],
            "title": "Action"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "string"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "action"
        ],
        "title": "ConflictResolutionRequest",
        "description": "Request body for POST /exports/conflicts/resolve.\n\nmode:\n  \"all\"       \u2013 apply action to every conflict in the project.\n  \"selection\" \u2013 apply action to the photos listed in `items`.\n\naction:\n  \"replace\"  \u2013 reset state to \"new\" so the photo is re-queued.\n  \"skip\"     \u2013 mark the photo as \"skip\" (do not process again).",
        "example": {
          "action": "replace",
          "items": [
            [
              "Session1",
              "IMG_001.jpg"
            ],
            [
              "Session1",
              "IMG_002.jpg"
            ]
          ],
          "mode": "selection"
        }
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard error envelope returned by exception handlers.",
        "example": {
          "detail": "Project 'my_project' not found."
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "example": {
          "password": "s3cr3t",
          "username": "admin"
        }
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access-token lifetime in seconds."
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "expires_in": 1800,
          "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "token_type": "bearer"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}